src/core/requirements.txt
alembic==1.3.2
certifi==2019.11.28
Flask==1.1.1
Flask-Cors==3.0.10
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.2
flask-oidc==1.4.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SSE==0.2.1
Flask-SQLAlchemy==2.4.4
gevent==21.8.0
greenlet==1.1.1
gunicorn==20.0.4
idna==2.9
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.5.0
marshmallow-enum==1.5.1
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-keycloak==0.23.0
pytz==2019.3
requests==2.26.0
schedule==0.6.0
six==1.13.0
sseclient-py==1.7
soupsieve==1.9.5
urllib3==1.26.7
Werkzeug==0.16.0
